copy-paste-fetaure/backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
gunicorn==20.1.0
Pillow==9.3.0  # Required for handling image processing

